requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
motor
PyJWT
pydantic[email]
cryptography
bson


